+++ noodles-tabix/src/reader.rs
@@ -49,7 +49,8 @@ where
             .set_reference_sequence_name_index(col_seq as usize)
             .set_start_position_index(col_beg as usize)
             .set_end_position_index(col_end as usize)
-            .set_comment(meta)
+            // Truncate comment line prefix.
+            .set_comment(meta as u8)
             .set_header_line_count(skip as u32)
             .set_reference_sequence_names(names)
             .set_reference_sequences(references);
